Refactoring solution.
