+++ src/main/java/org/example/epam/jat/volha/ResumeSearch.java
@@ -1,0 +1,10 @@
+package org.example.epam.jat.volha;
+
+import org.example.epam.jat.volha.data.Resume;
+
+public class ResumeSearch {
+  private Resume resume;
+  public void searchResume() {
+    //logic to search resume from some repo goes here
+  }
+}
