+++ src/main/java/org/example/epam/jat/volha/data/Resume.java
@@ -1,0 +1,9 @@
+package org.example.epam.jat.volha.data;
+
+public class Resume {
+  private int candidateId;
+  private String candidateNameAndSurname;
+  private Technology technology;
+  private Integer yearsOfExperience;
+
+}
