+++ src/main/java/org/example/epam/jat/volha/data/Technology.java
@@ -1,0 +1,4 @@
+package org.example.epam.jat.volha.data;
+
+public enum Technology {
+}
